+++ src/main/java/com/kang/blog/repository/ReplyRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface ReplyRepository extends JpaRepository<Reply, Integer> {
+public interface ReplyRepository extends JpaRepository<Reply, Long> {
 
     @Modifying
     @Query(value="INSERT INTO reply(userId, boardId, content, createDate) VALUES(?,?,?,now())", nativeQuery = true)
-    void commentSave(int userId, int boardId, String content);
+    void commentSave(Long userId, Long boardId, String content);
 }
